llm_multimodal/requirements.txt
transformers
whisper
pillow
ollama